move shared test fixtures to utils
